Move CommandPermission into ApplicationCommandUpdatePermissionsEvent and rename ApplicationCommandPermission to CommandPermission
